# pb/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "service.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/compression/api/v1/image/compress": {
      "post": {
        "operationId": "CompressImage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/compression_serviceCompressImageResponse"
            }
          }
        },
        "tags": [
          "Compression"
        ]
      }
    },
    "/compression/api/v1/image/decompress": {
      "post": {
        "operationId": "DecompressImage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/compression_serviceDecompressImageResponse"
            }
          }
        },
        "tags": [
          "Compression"
        ]
      }
    },
    "/compression/api/v1/text/compress": {
      "get": {
        "operationId": "CompressText",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/compression_serviceCompressTextResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "byte"
          }
        ],
        "tags": [
          "Compression"
        ]
      }
    },
    "/compression/api/v1/text/decompress": {
      "get": {
        "operationId": "DecompressText",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/compression_serviceDecompressTextResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "byte"
          }
        ],
        "tags": [
          "Compression"
        ]
      }
    }
  },
  "definitions": {
    "compression_serviceCompressImageResponse": {
      "type": "object",
      "properties": {
        "response_code": {
          "type": "integer",
          "format": "int32"
        },
        "response_message": {
          "type": "string"
        },
        "image": {
          "type": "string",
          "format": "byte"
        },
        "format": {
          "type": "string"
        },
        "level": {
          "$ref": "#/definitions/compression_serviceCompressionLevel"
        }
      }
    },
    "compression_serviceCompressTextResponse": {
      "type": "object",
      "properties": {
        "response_code": {
          "type": "integer",
          "format": "int32"
        },
        "response_message": {
          "type": "string"
        },
        "text": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "compression_serviceCompressionLevel": {
      "type": "string",
      "enum": [
        "DefaultCompression",
        "NoCompression",
        "BestSpeed",
        "BestCompression"
      ],
      "default": "DefaultCompression"
    },
    "compression_serviceDecompressImageResponse": {
      "type": "object",
      "properties": {
        "response_code": {
          "type": "integer",
          "format": "int32"
        },
        "response_message": {
          "type": "string"
        },
        "image": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "compression_serviceDecompressTextResponse": {
      "type": "object",
      "properties": {
        "response_code": {
          "type": "integer",
          "format": "int32"
        },
        "response_message": {
          "type": "string"
        },
        "text": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "compression_serviceImageType": {
      "type": "string",
      "enum": [
        "PNG",
        "JPEG"
      ],
      "default": "PNG"
    }
  }
}